Add Heikin-Ashi bar type with nanosecond time resolution
## Summary
- Add `HeikinAshiBarPolicy` with per-symbol state tracking via `SymbolStateMap`
- Upgrade time-based policies to nanosecond resolution (supports µs/ms/s/min/hour intervals)
- Change `BarEvent.barTypeParam` and `BarPolicy::param()` from `uint32_t` to `uint64_t`
- Clean up legacy Candle documentation, replace with Bar API docs

## Changes
- New `HeikinAshiBarPolicy` implementing smoothed OHLC formula
- `TimeBarPolicy`/`HeikinAshiBarPolicy` now accept any `std::chrono::duration`
- Added 6 new tests (5 Heikin-Ashi + 1 sub-second intervals)
- Created `bar.md`, `bar_event.md`, `bar_bus.md` API documentation
- Removed obsolete `candle_*.md` docs

+++ include/flox/aggregator/aggregation_policy.h
@@ -24,7 +24,7 @@ concept BarPolicy = requires(T& policy, const TradeEvent& trade, Bar& bar, TimeP
   { policy.update(trade, bar) } noexcept -> std::same_as<void>;
   { policy.initBar(trade, bar) } noexcept -> std::same_as<void>;
   { T::kBarType } -> std::convertible_to<BarType>;
-  { policy.param() } noexcept -> std::convertible_to<uint32_t>;
+  { policy.param() } noexcept -> std::convertible_to<uint64_t>;
 };
 
 // Common OHLCV update - free function, inlined
